servers-open-port-tester-metasploit
This folder contains a collection of discovered IP addresses and associated links, which have been tested for open ports using the Metasploit Framework. It provides valuable information about the security of remote servers, listing accessible ports and services found on various hosts. This file is primarily used for penetration testing and network security assessments. It helps identify exposed services and potential vulnerabilities that could be exploited. The discovered IPs and links are crucial for further analysis and security improvements on the tested systems.